% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.cluster}
\alias{Front.cluster}
\title{Group selected frontal points into 8-connected regions}
\usage{
Front.cluster(points)
}
\arguments{
\item{points}{- GSDF field marking frontal grid points (non NA=frontal)}
}
\value{
A similar field with each point numbered according to its cluster
}
\description{
Each connected region is a single front.
}
