% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.cluster.to.linear}
\alias{Front.cluster.to.linear}
\title{Convert frontal clusters into linear features}
\usage{
Front.cluster.to.linear(clusters)
}
\arguments{
\item{clusters}{- GSDF of clusters - output from \code{\link{Front.cluster}}}
}
\value{
A list of GSDF.front
}
\description{
Takes the point of max longitude at each latitude
}
\details{
Clusters that span a single grid-cell of latitude are deleted.
}
