% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.find}
\alias{Front.find}
\title{Find fronts from 6-hr changes in the wind field}
\usage{
Front.find(uwnd, vwnd, uwnd.old, vwnd.old)
}
\arguments{
\item{uwnd}{- Current U wind field.}

\item{vwnd}{- Current V wind field.}

\item{uwnd.old}{- U wind field from 6-hours ago}

\item{vwnd.old}{- V wind field from 6-hours ago}
}
\value{
A list of GSDF.Front structures
}
\description{
Uses the method of Simmonds et al. 'Identification and Climatology of
  Southern Hemisphere Mobile Fronts in a Modern Reanalysis'
  JCLI, 2011.
}
\details{
Basically identifies regions where wind changes direction and speed
 over 6 hours, and fits a smooth line to them.
}
