% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.find.wind.change.points}
\alias{Front.find.wind.change.points}
\title{Find regions with suitable 6-hr changes in the wind field}
\usage{
Front.find.wind.change.points(uwnd, vwnd, uwnd.old, vwnd.old)
}
\arguments{
\item{uwnd}{- Current U wind field.}

\item{vwnd}{- Current V wind field.}

\item{uwnd.old}{- U wind field from 6-hours ago}

\item{vwnd.old}{- V wind field from 6-hours ago}
}
\value{
A list of GSDF.Front structures
}
\description{
Want NH regions where wind moves from SW quadrant to NW quadrant
 and meridional speeds change by at least 2m/s
and SH regions where wind moves from NW quadrant to SW quadrant
 and meridional speeds change by at least 2m/s
}
