% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.plot}
\alias{Front.plot}
\title{Plot a single front}
\usage{
Front.plot(front, gp, Options)
}
\arguments{
\item{front}{- single front structure}
}
\value{
Same front structures, smoothed.
}
\description{
Rotates into the pole specified in the options
 and adjusts for longitude wraparounds
}
\details{
Probably this should be in GSDF.WeatherMap.
}
