% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.smooth.longitudes}
\alias{Front.smooth.longitudes}
\title{Smooth a series of longitudes}
\usage{
Front.smooth.longitudes(lons)
}
\arguments{
\item{lons}{- vector of longitudes (at least 2)}
}
\value{
Same longitudes, smoothed.
}
\description{
Uses the 5353H smoother, twice
}
