% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{Front.smooth}
\alias{Front.smooth}
\title{Smooth the raw front longitudes}
\usage{
Front.smooth(fronts)
}
\arguments{
\item{fronts}{- list of front structures - output from \code{\link{Front.cluster.to.linear}}}
}
\value{
Same front structures, smoothed.
}
\description{
Uses the 5353H smoother, twice
}
