% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.Front.R
\name{GSDF.Front}
\alias{GSDF.Front}
\title{Data structure describing a weather front}
\usage{
GSDF.Front()
}
\value{
A list as described (all components empty)
}
\description{
Fundamentally just a list with particular components:
\describe{
 \item{$lat} {numeric vector of latitudes}
 \item{$lon} {numeric vector of longitudes}
 \item{$type} {'cold','warm','stationary', or 'occluded'.
              NULL if unknown}
 \item{$moving} {'east' or 'west', NULL if unknown}
 \item{$meta} {any other info (list - key-value pairs)} Notably:
      pole.lat - latitude of the pole used (assume 90 if unspecified)
      pole.lon - longitude of the pole used (assume 180 if unspecified)
}
}
